VendingMachine Capstone Project
Created a vending machine simulator that will display items that are currently inside the vending machine and what is available to purchase and will allow the user to purchase multiple items until quantity is zero. Displays the changed given to the user after the end of purchasing items and producing a log file that will show all the actions from the vending machine such as taking in money and giving change to showing when and what was purchased.